requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
click==8.1.7
Django==5.1.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-tinymce==4.1.0
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
psycopg2-binary==2.9.10
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
webencodings==0.5.1
Werkzeug==3.0.6
